requirements.txt
hassil==3.0.1
PyYAML==6.0.2
voluptuous==0.15.2
regex==2024.11.6
jinja2==3.1.6

black==25.1.0
flake8==7.2.0
isort==6.0.1
mypy==1.15.0
pylint==3.3.7
pytest==8.3.5
pytest-xdist[psutil]==3.6.1
